refactor and cleanup
