+++ frontend/app/layout.js
@@ -1,5 +1,6 @@
 import './globals.css';
 import Link from 'next/link';
+import CartBadge from '../components/CartBadge';
 
 export const metadata = {
   title: 'Marcus Sports Shop',
@@ -10,7 +11,6 @@ export default function RootLayout({ children }) {
   return (
     <html lang="en">
       <body className="min-h-screen">
-        {/* Navigation */}
         <nav className="bg-white shadow-sm border-b border-gray-200">
           <div className="max-w-7xl mx-auto px-4">
             <div className="flex justify-between items-center h-16">
@@ -22,15 +22,17 @@ export default function RootLayout({ children }) {
                 <Link href="/" className="text-gray-600 hover:text-gray-900 transition-colors">
                   Products
                 </Link>
-                <Link href="/cart" className="text-gray-600 hover:text-gray-900 transition-colors">
-                  Cart
-                </Link>
+                <div className="flex items-center space-x-1">
+                  <Link href="/cart" className="text-gray-600 hover:text-gray-900 transition-colors">
+                    Cart
+                  </Link>
+                  <CartBadge />
+                </div>
               </div>
             </div>
           </div>
         </nav>
 
-        {/* Main Content */}
         <main className="bg-gray-50">
           {children}
         </main>
