+++ frontend/app/page.js
@@ -1,4 +1,4 @@
-import ProductList from './components/ProductList';
+import ProductList from '../components/ProductList';
 
 export default function Home() {
   return <ProductList />;
